comenzando a proteger

+++ srcs/bin/env_unset.c
@@ -6,7 +6,7 @@
 /*   By: tvillare <tvillare@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/09 13:17:45 by tvillare          #+#    #+#             */
-/*   Updated: 2023/04/14 18:47:36 by tvillare         ###   ########.fr       */
+/*   Updated: 2023/04/14 18:59:47 by tvillare         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -68,6 +68,8 @@ char	**unset_env(char **env, char **comand)
 		return (env);
 	len = len_doble_base(env);
 	new_env = ft_calloc((len - del) + 1, sizeof(char *));
+	if (new_env == NULL)
+		exit (1);
 	delete_unset(new_env, env, comand);
 	free (env);
 	int tmp;
